install_gitlab {remotes}	R Documentation

Install a package from GitLab

Description

This function is vectorised on repo so you can install multiple packages in a single command. Like other remotes the repository will skip installation if force == FALSE (the default) and the remote state has not changed since the previous installation.

Usage

install_gitlab(
  repo,
  subdir = NULL,
  auth_token = gitlab_pat(quiet),
  host = "gitlab.com",
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  repos = getOption("repos"),
  type = getOption("pkgType"),
  ...
)

Arguments

repo
Repository address in the format username/repo[@ref].

subdir
Subdirectory within repo that contains the R package.

auth_token
To install from a private repo, generate a personal access token (PAT) with at least read_api scope in https://docs.gitlab.com/ee/user/profile/personal_access_tokens.html and supply to this argument. This is safer than using a password because you can easily delete a PAT without affecting any others. Defaults to the GITLAB_PAT environment variable.

host
GitLab API host to use. Override with your GitLab enterprise hostname, for example, "<PROTOCOL://>gitlab.hostname.com". The PROTOCOL is required by packrat during RStudio Connect deployment. While install_gitlab may work without, omitting it generally leads to package restoration errors.

dependencies
Which dependencies do you want to check? Can be a character vector (selecting from "Depends", "Imports", "LinkingTo", "Suggests", or "Enhances"), or a logical vector.

TRUE is shorthand for "Depends", "Imports", "LinkingTo" and "Suggests". NA is shorthand for "Depends", "Imports" and "LinkingTo" and is the default. FALSE is shorthand for no dependencies (i.e. just check this package, not its dependencies).

The value "soft" means the same as TRUE, "hard" means the same as NA.

You can also specify dependencies from one or more additional fields, common ones include:

Config/Needs/website - for dependencies used in building the pkgdown site.

Config/Needs/coverage for dependencies used in calculating test coverage.

upgrade
Should package dependencies be upgraded? One of "default", "ask", "always", or "never". "default" respects the value of the R_REMOTES_UPGRADE environment variable if set, and falls back to "ask" if unset. "ask" prompts the user for which out of date packages to upgrade. For non-interactive sessions "ask" is equivalent to "always". TRUE and FALSE are also accepted and correspond to "always" and "never" respectively.

force
Force installation, even if the remote state has not changed since the previous install.

quiet
If TRUE, suppress output.

build
If TRUE build the package before installing.

build_opts
Options to pass to R CMD build, only used when build is TRUE.

build_manual
If FALSE, don't build PDF manual ('–no-manual').

build_vignettes
If FALSE, don't build package vignettes ('–no-build-vignettes').

repos
A character vector giving repositories to use.

type
Type of package to update.

...
Other arguments passed on to utils::install.packages().

See Also

Other package installation: install_bioc(), install_bitbucket(), install_cran(), install_dev(), install_github(), install_git(), install_local(), install_svn(), install_url(), install_version()

Examples

## Not run: 
install_gitlab("jimhester/covr")

## End(Not run)

[Package remotes version 2.4.2 Index]